Move `AnyT` to seperate module
