+++ rust/epac-flatbuffer-formats-rs/Cargo.toml
@@ -13,6 +13,7 @@ pyo3 = { version = "0.29.0", features = ["extension-module", "abi3-py38", "exper
 numpy = "0.29.0"
 flatbuffers = "=23.5.26"
 epac_flatbuffers_derive = { path = "../epac_flatbuffers_derive"}
+paste = "=1.0.15"
 
 [package.metadata.maturin]
 name = "epac_flatbuffer_formats_rs._bindings"
